refactor(orbit-visual): unify natal and transit into one shared component
Hoist the orbit visual into components/astrology/orbit-visual.tsx so the
natal birth-chart view and the transit/technical view share exactly one
implementation, styled like the transit visual: plain serif planet name +
mono degree text, no dignity color pills, no legend. The shared component
accepts `planets`, optional `dateLabel`, and optional `highlightPlanets`
for the purple-aura spotlight; Rahu/Ketu share a single outer ring (Ketu
mirrored).

The existing birth-chart/orbit-visual.tsx and chat/horoscope/transit-
orbit-visual.tsx files stay as thin adapters that pull the right slice
out of their respective data shapes, so all current call sites keep
working unchanged.
